+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 ### added
+* added feature flag `time`: Implements `Instant` for `time::Instant`.
+* added `TimeSw` type alias to `StopwatchImpl<time::Instant>`
 * added `StopwatchImpl::checked_toggle`
 * added `StopwatchImpl::checked_toggle_at`
 * added documentation notes about overflow to `StopwatchImpl::toggle` and `StopwatchImpl::toggle_at`